+++ Assets/Scripts/Test/TestMapController.cs
@@ -18,7 +18,7 @@ public class TestMapController : MapController, IInitializable
     [SerializeField]
     private ColorType _startBGColor;
     [SerializeField]
-    private float _mapSize;
+    private int _mapSize;
     public string Filename;
     private string _loadedFilename;
     private Stack<MapData> _moveRecord = new Stack<MapData>();
